+++ CHANGELOG.rst
@@ -4,7 +4,11 @@ CHANGELOG
 6.0.0a13 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- Add `max_ops` property to `PatchField`, `BucketListField` and `BucketDictField`
+  [vangheem]
+
+- Add clear action to list, dict and annotation patch fields
+  [vangheem]
 
 
 6.0.0a12 (2020-02-18)